Rolling Update part 1/2
Added a all task view

Change weekly tasks to separate save preference in order to learn more about potential save preference options.

ran in to a unknown error with app crashing when switching to the task page, while the problem was fixed in this update the reason behind it was unknown.

Part two will have delete selected options in all task view, and all tasks will probably be moved to edit tasks and expanded on at a later date.